devoir/requirements.txt
# Federated Learning with Kafka and Spark
# Python dependencies

# Kafka
kafka-python==2.0.2

# Spark
pyspark==3.5.0

# Scientific computing
numpy==1.26.2

# Utilities
python-dateutil==2.8.2
